+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+Version 9:
+
+* Added swap method to net::socket class
+* Added noexcept keyword to move constructor and operator of class net::socket
+* Fixed README.md
+
+--------------------------------------------------------------------------------
+
 Version 8:
 
 * Introduced tcp socket unit test